test: updated Docker compose GH workflow test
Added the required start_interval to support Docker Compose
versions > 2.36.0

Changed the container versions to be the latest compatible with node,
however I suspect this might be dangerous. As this is for local
development and testing, maybe we can provide container version defaults
that are compliant with the comp. matrix in testnet02, and these can be
overridden with env variables?

Signed-off-by: Giuseppe Salvatore <giuseppe.salvatore@shielded.io>